requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-genai>=1.0.0
pdfplumber==0.11.0
python-docx==1.1.2
sentence-transformers==2.7.0
faiss-cpu==1.13.0
numpy>=1.24.0
pydantic>=2.0